+++ example/lib/multi_scenario/maps/map_biome_1.dart
@@ -47,7 +47,7 @@ class MapBiome1 extends StatelessWidget {
       ),
       cameraConfig: CameraConfig(
         moveOnlyMapArea: true,
-        zoom: getZoomGame(context, defaultTileSize, 20),
+        zoom: getZoomFromMaxVisibleTile(context, defaultTileSize, 20),
       ),
       progress: Container(
         width: double.maxFinite,