+++ example/lib/shared/util/functions.dart
@@ -2,7 +2,11 @@ import 'dart:math';
 
 import 'package:flutter/widgets.dart';
 
-double getZoomGame(BuildContext context, double tileSize, int maxTile) {
+double getZoomFromMaxVisibleTile(
+  BuildContext context,
+  double tileSize,
+  int maxTile,
+) {
   final screenSize = MediaQuery.of(context).size;
   final maxSize = max(screenSize.width, screenSize.height);
   return maxSize / (tileSize * maxTile);