+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## NEXT
+- extension method `simpleAttackMeleeByDirection` improvements
+
 ## 3.0.0-beta.7
 
 - `MovementByJoystick` improvements. Adds `setupMovementByJoystick` method.